% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scIA_finest.R
\name{scIA_finest}
\alias{scIA_finest}
\title{Single-cell intelligent annotation (finest)}
\usage{
scIA_finest(df, pan_mapping = NULL)
}
\arguments{
\item{df}{A dataframe containing the semi-supervised and unsupervised results of scIA function}

\item{pan_mapping}{Mapping cells to main cell types}
}
\value{
A dataframe containing the final results
}
\description{
This function determines more accurate cell type through a complex decision system
based on the semi-supervised and unsupervised results of scIA function.
}
\examples{
\dontrun{
scIA_finest(df,
pan_mapping = pan_mapping
}
}
