% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scReType.R
\name{scReType_multi}
\alias{scReType_multi}
\title{Single-cell re-identification analysis (multi)
This function performs multi-cell re-identification by evaluating
cells against multiple candidate cell types simultaneously and assigning
the most probable label based on marker gene expression patterns.}
\usage{
scReType_multi(
  object,
  cell_ids,
  current_labels,
  compared_labels,
  gs,
  shared_gs = NULL,
  add_gs_exp = FALSE
)
}
\arguments{
\item{object}{Seurat object}

\item{cell_ids}{Name of cells to be analyzed}

\item{current_labels}{The current cell type of the cells}

\item{compared_labels}{The cell type to be identified}

\item{gs}{List of marker genes}

\item{shared_gs}{Shared marker genes for cell types to be identified}

\item{add_gs_exp}{Whether to print the gene expression in the results}
}
\value{
A dataframe containing the final results
}
\description{
Single-cell re-identification analysis (multi)
This function performs multi-cell re-identification by evaluating
cells against multiple candidate cell types simultaneously and assigning
the most probable label based on marker gene expression patterns.
}
\examples{
\dontrun{
scReType_multi(sce, 
cell_ids = cell_ids, 
current_labels = 'Undetermined',
compared_labels = 'Neutrophils',
gs = gs, 
shared_gs = shared_gs,
add_gs_exp = FALSE
}
}
