% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scReType.R
\name{scReType_single}
\alias{scReType_single}
\title{Single-cell re-identification analysis (single)
This function performs single-cell re-identification by evaluating
whether cells expressing a sufficient number of marker genes should
be reclassified to a target cell type.}
\usage{
scReType_single(
  object,
  cell_ids,
  current_labels,
  compared_labels,
  gs,
  threshold = 3,
  add_gs_exp = FALSE
)
}
\arguments{
\item{object}{Seurat object}

\item{cell_ids}{Name of cells to be analyzed}

\item{current_labels}{The current cell type of the cells}

\item{compared_labels}{The cell type to be identified}

\item{gs}{List of marker genes}

\item{threshold}{The number of marker genes}

\item{add_gs_exp}{Whether to print the gene expression in the results}
}
\value{
A dataframe containing the final results
}
\description{
Single-cell re-identification analysis (single)
This function performs single-cell re-identification by evaluating
whether cells expressing a sufficient number of marker genes should
be reclassified to a target cell type.
}
\examples{
\dontrun{
scReType_single(sce, 
cell_ids = cell_ids, 
current_labels = 'Undetermined',
compared_labels = 'Neutrophils',
gs = gs, 
threshold = 2,
add_gs_exp = FALSE
}
}
