% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scIA.R
\name{scIA}
\alias{scIA}
\title{Single-cell intelligent annotation}
\usage{
scIA(
  object,
  org = c("hsa", "mus", "rat"),
  cell_ids = NULL,
  ex_types = NULL,
  gs = NULL,
  ci_gs = NULL,
  wt_gs = NULL,
  pan_mapping = NULL,
  include_gs = FALSE,
  use_pan = FALSE,
  use_cluster = TRUE,
  cluster_ids = NULL,
  cluster_col = "seurat_clusters"
)
}
\arguments{
\item{object}{Seurat object.}

\item{org}{Organism}

\item{cell_ids}{Name of cells to be analyzed}

\item{ex_types}{Cell types to be excluded}

\item{gs}{List of gene markers}

\item{ci_gs}{List of gene markers' confidence}

\item{wt_gs}{List of gene markers' weight}

\item{pan_mapping}{Mapping cells to main cell types}

\item{include_gs}{Whether to print the gene expression in the results}

\item{use_pan}{Using pan-celltype mapping function}

\item{use_cluster}{Using cluster mapping function (Generated by FindClusters function)}

\item{cluster_ids}{Cluster ids of object (Generated by FindClusters function)}

\item{cluster_col}{The column name of cluster ids in metadata}
}
\value{
A dataframe containing the final results
}
\description{
Single-cell intelligent annotation
}
\examples{
scIA(sce, org = 'hsa', 
cell_ids = cell_ids, 
gs = gs, 
ex_types = c('Goblet (bronchial)','SMG duct','SMG mucous','SMG serous (bronchial)'), 
ci_gs = ci_gs,
wt_gs = wt_gs,
include_gs = FALSE,
pan_mapping = pan_celltype,
use_pan = FALSE,
use_cluster = TRUE,  
cluster_ids = cluster_ids,   
cluster_col = "seurat_clusters"
}
\author{
Fanjie Lin, Yuanfang Xie, Weiguo Kong, Ken Lin
\email{lfj@stu.gzhmu.edu.cn}
}
